Chỉnh sửa các entity và thêm mới các repository

+++ src/main/java/com/VietBlog/repository/BaiVietRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BaiVietRepository extends JpaRepository<BaiViet, Integer> {
+public interface BaiVietRepository extends JpaRepository<BaiViet, Long> {
     // Đếm tổng số lượng bài viết
     @Query("SELECT COUNT(b.id) FROM BaiViet b")
     Integer countTotalBaiViet();
